Update README.md
Rhetoricall is an open source, SMS based public engagement project that allows the public to send messages from their cellphones to a shared public display. The project is open source and invites programmers, designers, hackers to create their own visualizations for incoming text messages to add to a growing library of effects for visual output. Based on research done on my earlier projects, SimpleTEXT and Lively, Rhetoricall builds on the success of these projects but encourages collaboration by the public to enhance the projects even further.